Add new namespace called Integrations
This will be used for services such as Notify to webhook with us.
