+++ config/routes.rb
@@ -281,7 +281,9 @@ Rails.application.routes.draw do
   get '/auth/dfe/callback' => 'dfe_sign_in#callback'
   post '/auth/developer/callback' => 'dfe_sign_in#bypass_callback'
 
-  post '/notify/callback' => 'notify#callback'
+  namespace :integrations, path: '/integrations' do
+    post '/notify/callback' => 'notify#callback'
+  end
 
   namespace :support_interface, path: '/support' do
     get '/' => redirect('/support/candidates')
